requirements.txt
pandas
numpy
yfinance
psycopg2-binary
scikit-learn
xgboost
lightgbm
joblib
python-dotenv
sqlalchemy
matplotlib
seaborn
jupyter
